# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'AZE FARM API'
  version: 1.0.0
  description: 'API pour la gestion des produits agricoles'
servers:
  -
    url: 'http://localhost:3000/api'
    description: 'Serveur de développement'
paths:
  /api/products/search:
    get:
      tags:
        - Products
      summary: 'Rechercher des produits avec filtres et localisation'
      description: "Recherche des produits disponibles avec filtres avancés et tri par proximité géographique.\nLes produits sont classés par distance de l'utilisateur si les coordonnées sont fournies,\nsinon par note de la ferme puis par date de création.\n\n**Filtres disponibles** :\n- `farmId` : Filtrer par ferme spécifique\n- `seasonal=true` : Uniquement les produits de saison actuelle\n\n**Mode suggestions** : Ajoutez `suggestions=true` pour obtenir des suggestions de noms de produits.\nDans ce mode, seul le paramètre `product` est requis.\n"
      parameters:
        -
          in: query
          name: suggestions
          schema:
            type: boolean
            default: false
          description: 'Activer le mode suggestions (retourne uniquement des noms de produits)'
        -
          in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: 'Nombre maximum de résultats à retourner'
        -
          in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 'Numéro de page pour la pagination (ignoré en mode suggestions)'
        -
          in: query
          name: category
          schema:
            type: string
            example: légumes
          description: 'Filtrer par catégorie de produit (ignoré en mode suggestions)'
        -
          in: query
          name: product
          schema:
            type: string
            example: tomate
          description: "Filtrer par nom de produit (recherche partielle).\nEn mode suggestions, ce paramètre est requis pour générer les suggestions.\n"
        -
          in: query
          name: lat
          schema:
            type: number
            minimum: -90
            maximum: 90
            example: 48.8566
          description: 'Latitude de l''utilisateur pour le calcul de distance (ignoré en mode suggestions)'
        -
          in: query
          name: lng
          schema:
            type: number
            minimum: -180
            maximum: 180
            example: 2.3522
          description: 'Longitude de l''utilisateur pour le calcul de distance (ignoré en mode suggestions)'
        -
          in: query
          name: farmId
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
          description: 'Filtrer par ID de ferme spécifique (ignoré en mode suggestions)'
        -
          in: query
          name: seasonal
          schema:
            type: boolean
            default: false
          description: 'Uniquement les produits de saison actuelle (ignoré en mode suggestions)'
        -
          in: query
          name: userId
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
          description: 'Personnaliser les résultats selon les préférences utilisateur (ignoré en mode suggestions)'
        -
          in: query
          name: favorites
          schema:
            type: boolean
            default: false
          description: 'Mettre en avant les produits favoris de l''utilisateur (ignoré en mode suggestions)'
        -
          in: query
          name: history
          schema:
            type: boolean
            default: false
          description: 'Éviter les doublons avec l''historique de recherche (ignoré en mode suggestions)'
      responses:
        '200':
          description: 'Produits ou suggestions récupérés avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Produits récupérés avec succès'}
                  data: {oneOf: [{type: object, description: 'Mode recherche normale', properties: {items: {type: array, items: {$ref: '#/components/schemas/ProductWithDistance'}}, totalItems: {type: integer, example: 25}, totalPages: {type: integer, example: 3}, currentPage: {type: integer, example: 1}}}, {type: array, description: 'Mode suggestions', items: {type: string}, example: [tomate, 'tomate cerise', 'tomate cœur de bœuf']}]}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/api/products/{id}':
    get:
      tags:
        - Products
      summary: 'Obtenir les détails d''un produit'
      description: 'Récupère les informations complètes d''un produit spécifique incluant les détails de la ferme'
      parameters:
        -
          in: path
          name: id
          required: true
          schema:
            type: string
            format: mongoId
          description: 'ID du produit'
      responses:
        '200':
          description: 'Produit récupéré avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Produit récupéré avec succès'}
                  data: {type: object, properties: {name: {type: string, example: 'Tomate cerise'}, description: {type: string, example: 'Tomates cerise sucrées et juteuses'}, price: {type: number, example: 4.5}, unit: {type: string, example: kg}, stock: {type: integer, example: 25}, image: {type: string, example: 'https://example.com/image.jpg'}, farm: {type: object, properties: {name: {type: string, example: 'Ferme du Soleil'}, image: {type: string, example: 'https://example.com/farm.jpg'}, rating: {type: object, properties: {average: {type: number, example: 4.5}, count: {type: integer, example: 12}}}}}, otherProducts: {type: array, items: {type: object, properties: {id: {type: string}, name: {type: string}, price: {type: number}, unit: {type: string}, stock: {type: integer}, image: {type: string}, createdAt: {type: string, format: date-time}}}}}}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/products/categories:
    get:
      tags:
        - Products
      summary: 'Lister les catégories de produits'
      description: 'Récupère la liste de toutes les catégories de produits disponibles'
      responses:
        '200':
          description: 'Catégories récupérées avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Catégories récupérées avec succès'}
                  data: {type: array, items: {type: string}, example: [légumes, fruits, céréales, 'produits laitiers']}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/products/trends:
    get:
      tags:
        - Products
      summary: 'Obtenir les tendances de recherche'
      description: "Retourne les recherches les plus populaires avec leur croissance,\nainsi que des statistiques sur l'utilisation de la recherche.\n"
      parameters:
        -
          in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: 'Nombre de tendances à retourner'
        -
          in: query
          name: days
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 30
          description: 'Période d''analyse en jours'
        -
          in: query
          name: stats
          schema:
            type: boolean
            default: false
          description: 'Inclure les statistiques détaillées'
      responses:
        '200':
          description: 'Tendances récupérées avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Tendances récupérées avec succès'}
                  data: {type: object, properties: {trending: {type: array, items: {type: object, properties: {term: {type: string, example: tomate}, count: {type: integer, example: 45}, growth: {type: number, example: 23.5}}}}, stats: {type: object, properties: {totalSearches: {type: integer, example: 1250}, uniqueTerms: {type: integer, example: 89}, avgResponseTime: {type: integer, example: 120}, topCategories: {type: array, items: {type: object, properties: {category: {type: string}, count: {type: integer}}}}}}}}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/products/suggestions:
    get:
      tags:
        - Products
      summary: '[OBSOLÈTE] Obtenir des suggestions de produits'
      deprecated: true
      description: "**CET ENDPOINT EST OBSOLÈTE**\n\nUtilisez plutôt `/api/products/search?suggestions=true&product=votre_terme`\n\nRetourne des suggestions de produits basées sur un terme de recherche partiel\n"
      parameters:
        -
          in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 50
          description: 'Terme de recherche pour suggestions'
        -
          in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
          description: 'Nombre de suggestions à retourner'
      responses:
        '200':
          description: 'Suggestions récupérées avec succès'
        '400':
          description: 'Erreur de validation'
        '500':
          description: 'Erreur serveur'
...